Resolved a set of warnings when compiling with MSVC x64 for: * gl2ps * chi2mcrl2 * grapemcrl2
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8561 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ tools/chi2mcrl2/translate.cpp
@@ -251,7 +251,7 @@ std::string CAsttransform::manipulateProcess(ATermAppl input)
   vector<RAT>::iterator itRAT;
   vector<int>::iterator itint;
   string result;
-  set<int> collection_of_used_counters;
+  set<size_t> collection_of_used_counters;
 
   //Clear global variables that are exclusivly thoughout this process
   begin_state.clear(); //first:  parenthesis level
